dictation/requirements_windows.txt
pystray
Pillow
pynput
sounddevice
soundfile
numpy
requests
pyperclip
webrtcvad-wheels
